[DBC] Don't die so easily on UTF-8 decoding errors on strings
Instead, check if the final byte of the string is a start-of multi byte
codepoint, and if so, try to decode again without the last byte. Note
that if this decode also fails (results in an exception being thrown),
the program will exit.

In other cases, return no string and print out the decoding error to the
user.